Move topbar menu items into new component
