+++ src/components/TopBar/TopBar.vue
@@ -65,6 +65,7 @@
 		<!-- TopBar menu -->
 		<TopBarMenu :token="token"
 			:show-actions="!isSidebar"
+			:is-sidebar="isSidebar"
 			:model="localMediaModel" />
 
 		<div class="top-bar__buttons">
@@ -75,63 +76,6 @@
 				class="top-bar__separator" />
 
 			<!-- sidebar toggle -->
-			<NcActions v-if="!isSidebar"
-				v-shortkey.once="disableKeyboardShortcuts ? null : ['f']"
-				class="top-bar__button"
-				:aria-label="t('spreed', 'Conversation actions')"
-				:container="container"
-				@shortkey.native="toggleFullscreen">
-				<template #icon>
-					<span :class="{'top-bar__button__force-white': isInCall}">
-						<Cog :size="20" />
-					</span>
-				</template>
-				<NcActionButton :icon="iconFullscreen"
-					:aria-label="t('spreed', 'Toggle fullscreen')"
-					:close-after-click="true"
-					@click="toggleFullscreen">
-					{{ labelFullscreen }}
-				</NcActionButton>
-				<NcActionSeparator v-if="showModerationOptions" />
-				<NcActionLink v-if="isFileConversation"
-					:href="linkToFile">
-					<template #icon>
-						<File :size="20" />
-					</template>
-					{{ t('spreed', 'Go to file') }}
-				</NcActionLink>
-				<template v-if="showModerationOptions">
-					<NcActionButton :close-after-click="true"
-						icon="icon-rename"
-						@click="handleRenameConversation">
-						{{ t('spreed', 'Rename conversation') }}
-					</NcActionButton>
-				</template>
-				<NcActionButton v-if="!isOneToOneConversation"
-					icon="icon-clippy"
-					:close-after-click="true"
-					@click="handleCopyLink">
-					{{ t('spreed', 'Copy link') }}
-				</NcActionButton>
-				<template v-if="showModerationOptions && canFullModerate && isInCall">
-					<NcActionSeparator />
-					<NcActionButton :close-after-click="true"
-						@click="forceMuteOthers">
-						<template #icon>
-							<MicrophoneOff :size="20" />
-						</template>
-						{{ t('spreed', 'Mute others') }}
-					</NcActionButton>
-				</template>
-				<NcActionSeparator v-if="showModerationOptions" />
-				<NcActionButton :close-after-click="true"
-					@click="openConversationSettings">
-					<template #icon>
-						<Cog :size="20" />
-					</template>
-					{{ t('spreed', 'Conversation settings') }}
-				</NcActionButton>
-			</NcActions>
 			<NcActions v-if="showOpenSidebarButton"
 				class="top-bar__button"
 				close-after-click="true"
@@ -162,28 +106,23 @@
 </template>
 
 <script>
-import { showError, showSuccess, showMessage } from '@nextcloud/dialogs'
+import { showMessage } from '@nextcloud/dialogs'
 import NcActionButton from '@nextcloud/vue/dist/Components/NcActionButton.js'
 import NcActions from '@nextcloud/vue/dist/Components/NcActions.js'
 import NcCounterBubble from '@nextcloud/vue/dist/Components/NcCounterBubble.js'
 import CallButton from './CallButton.vue'
 import BrowserStorage from '../../services/BrowserStorage.js'
-import NcActionLink from '@nextcloud/vue/dist/Components/NcActionLink.js'
-import NcActionSeparator from '@nextcloud/vue/dist/Components/NcActionSeparator.js'
-import File from 'vue-material-design-icons/File.vue'
 import MenuPeople from '../missingMaterialDesignIcons/MenuPeople.vue'
 import MessageText from 'vue-material-design-icons/MessageText.vue'
-import MicrophoneOff from 'vue-material-design-icons/MicrophoneOff.vue'
 import { CONVERSATION, PARTICIPANT } from '../../constants.js'
 import { generateUrl } from '@nextcloud/router'
-import { callParticipantCollection, localCallParticipantModel, localMediaModel } from '../../utils/webrtc/index.js'
+import { localCallParticipantModel, localMediaModel } from '../../utils/webrtc/index.js'
 import { emit } from '@nextcloud/event-bus'
 import ConversationIcon from '../ConversationIcon.vue'
 import Tooltip from '@nextcloud/vue/dist/Directives/Tooltip.js'
 import richEditor from '@nextcloud/vue/dist/Mixins/richEditor.js'
 import userStatus from '../../mixins/userStatus.js'
 import LocalMediaControls from '../CallView/shared/LocalMediaControls.vue'
-import Cog from 'vue-material-design-icons/Cog.vue'
 import getParticipants from '../../mixins/getParticipants.js'
 import TopBarMenu from './TopBarMenu.vue'
 
@@ -197,17 +136,12 @@ export default {
 	components: {
 		NcActionButton,
 		NcActions,
-		NcActionLink,
 		NcCounterBubble,
 		CallButton,
-		NcActionSeparator,
-		File,
 		MenuPeople,
 		MessageText,
-		MicrophoneOff,
 		ConversationIcon,
 		LocalMediaControls,
-		Cog,
 		TopBarMenu,
 	},
 
@@ -246,58 +180,16 @@ export default {
 			return this.$store.getters.getMainContainerSelector()
 		},
 
-		isFullscreen() {
-			return this.$store.getters.isFullscreen()
-		},
-
-		iconFullscreen() {
-			if (this.isInCall) {
-				return 'forced-white icon-fullscreen'
-			}
-			return 'icon-fullscreen'
-		},
-
-		labelFullscreen() {
-			if (this.isFullscreen) {
-				return t('spreed', 'Exit fullscreen (F)')
-			}
-			return t('spreed', 'Fullscreen (F)')
-		},
-
 		showOpenSidebarButton() {
 			return !this.$store.getters.getSidebarStatus
-		},
-
-		isFileConversation() {
-			return this.conversation.objectType === 'file' && this.conversation.objectId
 		},
 
 		isOneToOneConversation() {
 			return this.conversation.type === CONVERSATION.TYPE.ONE_TO_ONE
 		},
 
-		linkToFile() {
-			if (this.isFileConversation) {
-				return window.location.protocol + '//' + window.location.host + generateUrl('/f/' + this.conversation.objectId)
-			} else {
-				return ''
-			}
-		},
-
 		participantType() {
 			return this.conversation.participantType
-		},
-
-		canFullModerate() {
-			return this.participantType === PARTICIPANT.TYPE.OWNER || this.participantType === PARTICIPANT.TYPE.MODERATOR
-		},
-
-		canModerate() {
-			return this.canFullModerate || this.participantType === PARTICIPANT.TYPE.GUEST_MODERATOR
-		},
-
-		showModerationOptions() {
-			return !this.isOneToOneConversation && this.canModerate
 		},
 
 		token() {
@@ -329,11 +221,6 @@ export default {
 			} else {
 				return ''
 			}
-		},
-
-		conversationHasSettings() {
-			return this.conversation.type === CONVERSATION.TYPE.GROUP
-				|| this.conversation.type === CONVERSATION.TYPE.PUBLIC
 		},
 
 		renderedDescription() {
@@ -368,10 +255,6 @@ export default {
 			if (peer) {
 				return !peer.sessionIds.length
 			} else return false
-		},
-
-		disableKeyboardShortcuts() {
-			return OCP.Accessibility.disableKeyboardShortcuts()
 		},
 	},
 
@@ -445,67 +328,6 @@ export default {
 		openSidebar() {
 			this.$store.dispatch('showSidebar')
 			BrowserStorage.setItem('sidebarOpen', 'true')
-		},
-
-		fullScreenChanged() {
-			this.$store.dispatch(
-				'setIsFullscreen',
-				document.webkitIsFullScreen || document.mozFullScreen || document.msFullscreenElement
-			)
-		},
-
-		toggleFullscreen() {
-			if (this.isFullscreen) {
-				this.disableFullscreen()
-				this.$store.dispatch('setIsFullscreen', false)
-			} else {
-				this.enableFullscreen()
-				this.$store.dispatch('setIsFullscreen', true)
-			}
-		},
-
-		enableFullscreen() {
-			const fullscreenElem = document.getElementById('content-vue')
-
-			if (fullscreenElem.requestFullscreen) {
-				fullscreenElem.requestFullscreen()
-			} else if (fullscreenElem.webkitRequestFullscreen) {
-				fullscreenElem.webkitRequestFullscreen(Element.ALLOW_KEYBOARD_INPUT)
-			} else if (fullscreenElem.mozRequestFullScreen) {
-				fullscreenElem.mozRequestFullScreen()
-			} else if (fullscreenElem.msRequestFullscreen) {
-				fullscreenElem.msRequestFullscreen()
-			}
-		},
-
-		disableFullscreen() {
-			if (document.exitFullscreen) {
-				document.exitFullscreen()
-			} else if (document.webkitExitFullscreen) {
-				document.webkitExitFullscreen()
-			} else if (document.mozCancelFullScreen) {
-				document.mozCancelFullScreen()
-			} else if (document.msExitFullscreen) {
-				document.msExitFullscreen()
-			}
-		},
-
-		async handleCopyLink() {
-			try {
-				await navigator.clipboard.writeText(this.linkToConversation)
-				showSuccess(t('spreed', 'Conversation link copied to clipboard'))
-			} catch (error) {
-				showError(t('spreed', 'The link could not be copied'))
-			}
-		},
-		handleRenameConversation() {
-			this.$store.dispatch('isRenamingConversation', true)
-			this.$store.dispatch('showSidebar')
-		},
-		forceMuteOthers() {
-			callParticipantCollection.callParticipantModels.forEach(callParticipantModel => {
-				callParticipantModel.forceMute()
-			})
 		},
 
 		openConversationSettings() {
